Revise logics of runtime CA detection
If CMake is used, it is very likely that rTorrent is being linked
against a shared cURL library from system. In this case, distro config
should be respected. As such, disable the feature on CMake build by
default.

Search from the highest priority first and always observe RTORRENT_CA_BUNDLE.