control: add an "urgent shutdown" mechanism
Pressure tests show that with 16000+ torrents, 60000+ open files
and 5000+ sockets, the shutdown process takes forever.

Service managers like systemd usually have a timeout. For instance,
by default, systemd SIGKILL the process after 90 seconds.

This change adds an "urgent shutdown" mechanism to save the session
and ensure that the shutdown is as graceful as it can be. It can be
further extended in the future.