download_list: allow to preallocate space for the whole torrent at once
Currently, with system.file.allocate = 1, rTorrent preallocates space
only after the file is opened for write.

In many cases, it is preferable to preallocate space for the whole
torrent at once, to reduce fragmentation of contents inside the
torrent.

It allows more efficient filesystem organization, where files that
are likely to be read together are packed in contiguous area of the
disk, improving the performance and I/O patterns of future read
operations.

This change, along with a change to libTorrent, implements the function:

system.file.allocate = 2