rpc: handle XML-RPC with JSON-RPC logics
This allows us to centralize on a single set of RPC logics, sharing
the exception handling, error codes/messages and behaviors.

This improves consistency, reduces maintenance burden, makes future
extension to RPC parts easier. In addition, we can ditch the dependency
on decade-old xmlrpc-c library.

As for the performance, we performed a test with 10,000 torrents in
a typical d.multicall2 call, which showed JSON-RPC is 3x faster,
compared to XML-RPC (xmlrpc-c). Meanwhile, rapidxml is 3-9x faster,
compared to expat/libxml. That means even with the conversion, we
shouldn't see much of a performance ding. We may even have a win.